+++ crates/ruma-client-api/src/delayed_events/delayed_message_event.rs
@@ -14,7 +14,7 @@ pub mod unstable {
     //! [MSC]: https://github.com/matrix-org/matrix-spec-proposals/pull/4140
 
     use ruma_common::{
-        OwnedRoomId, OwnedTransactionId,
+        OwnedTransactionId, RoomId,
         api::{auth_scheme::AccessToken, request, response},
         metadata,
         serde::Raw,
@@ -39,7 +39,7 @@ pub mod unstable {
     pub struct Request {
         /// The room to send the event to.
         #[ruma_api(path)]
-        pub room_id: OwnedRoomId,
+        pub room_id: RoomId,
 
         /// The type of event to send.
         #[ruma_api(path)]
@@ -83,7 +83,7 @@ pub mod unstable {
         /// Since `Request` stores the request body in serialized form, this function can fail if
         /// `T`s [`::serde::Serialize`] implementation can fail.
         pub fn new<T>(
-            room_id: OwnedRoomId,
+            room_id: RoomId,
             txn_id: OwnedTransactionId,
             delay_parameters: DelayParameters,
             content: &T,
@@ -103,7 +103,7 @@ pub mod unstable {
         /// Creates a new `Request` with the given room id, transaction id, event type,
         /// `delay_parameters` and raw event content.
         pub fn new_raw(
-            room_id: OwnedRoomId,
+            room_id: RoomId,
             txn_id: OwnedTransactionId,
             event_type: MessageLikeEventType,
             delay_parameters: DelayParameters,
@@ -130,7 +130,7 @@ pub mod unstable {
                 MatrixVersion, OutgoingRequestExt as _, SupportedVersions,
                 auth_scheme::SendAccessToken,
             },
-            owned_room_id,
+            room_id,
         };
         use ruma_events::room::message::RoomMessageEventContent;
         use serde_json::{Value as JsonValue, json};
@@ -141,7 +141,7 @@ pub mod unstable {
 
         #[test]
         fn serialize_delayed_message_request() {
-            let room_id = owned_room_id!("!roomid:example.org");
+            let room_id = room_id!("!roomid:example.org");
             let supported = SupportedVersions {
                 versions: [MatrixVersion::V1_1].into(),
                 features: Default::default(),