+++ crates/ruma-client-api/src/delayed_events/get_delayed_event.rs
@@ -67,7 +67,7 @@ pub mod unstable {
 
         use js_int::UInt;
         use ruma_common::{
-            MilliSecondsSinceUnixEpoch, api::OutgoingResponseExt as _, event_id, owned_room_id,
+            MilliSecondsSinceUnixEpoch, api::OutgoingResponseExt as _, event_id, room_id,
             serde::Raw,
         };
         use ruma_events::TimelineEventType;
@@ -85,7 +85,7 @@ pub mod unstable {
 
             let mut event_data = DelayedEventData::new(
                 "a_delay_id".to_owned(),
-                owned_room_id!("!roomid:example.org"),
+                room_id!("!roomid:example.org"),
                 TimelineEventType::RoomTopic,
                 Some("a_state_key".to_owned()),
                 Raw::from_json_string(content).unwrap(),
@@ -123,7 +123,7 @@ pub mod unstable {
 
         use js_int::UInt;
         use ruma_common::{
-            MilliSecondsSinceUnixEpoch, api::IncomingResponseExt as _, event_id, owned_room_id,
+            MilliSecondsSinceUnixEpoch, api::IncomingResponseExt as _, event_id, room_id,
         };
         use ruma_events::TimelineEventType;
         use serde_json::{Value as JsonValue, json};
@@ -158,7 +158,7 @@ pub mod unstable {
             });
 
             assert_eq!(res.delay_id, "a_delay_id".to_owned());
-            assert_eq!(res.room_id, owned_room_id!("!roomid:example.org"));
+            assert_eq!(res.room_id, room_id!("!roomid:example.org"));
             assert_eq!(res.event_type, TimelineEventType::RoomTopic);
             assert_eq!(res.state_key, Some("a_state_key".to_owned()));
             assert_eq!(res.delay, Duration::from_millis(103));