+++ crates/ruma-client-api/src/room/upgrade_room.rs
@@ -8,7 +8,7 @@ pub mod v3 {
     //! [spec]: https://spec.matrix.org/v1.19/client-server-api/#post_matrixclientv3roomsroomidupgrade
 
     use ruma_common::{
-        OwnedRoomId, OwnedUserId, RoomVersionId,
+        OwnedUserId, RoomId, RoomVersionId,
         api::{auth_scheme::AccessToken, request, response},
         metadata,
     };
@@ -33,7 +33,7 @@ pub mod v3 {
 
         /// ID of the room to be upgraded.
         #[ruma_api(path)]
-        pub room_id: OwnedRoomId,
+        pub room_id: RoomId,
 
         /// New version for the room.
         pub new_version: RoomVersionId,
@@ -43,19 +43,19 @@ pub mod v3 {
     #[response]
     pub struct Response {
         /// ID of the new room.
-        pub replacement_room: OwnedRoomId,
+        pub replacement_room: RoomId,
     }
 
     impl Request {
         /// Creates a new `Request` with the given room ID and new room version.
-        pub fn new(room_id: OwnedRoomId, new_version: RoomVersionId) -> Self {
+        pub fn new(room_id: RoomId, new_version: RoomVersionId) -> Self {
             Self { room_id, new_version, additional_creators: Vec::new() }
         }
     }
 
     impl Response {
         /// Creates a new `Response` with the given room ID.
-        pub fn new(replacement_room: OwnedRoomId) -> Self {
+        pub fn new(replacement_room: RoomId) -> Self {
             Self { replacement_room }
         }
     }