+++ crates/ruma-common/src/identifiers.rs
@@ -34,7 +34,7 @@ pub use self::{
     mxc_uri::MxcUri,
     one_time_key_name::OneTimeKeyName,
     room_alias_id::RoomAliasId,
-    room_id::{OwnedRoomId, RoomId},
+    room_id::RoomId,
     room_or_alias_id::{OwnedRoomOrAliasId, RoomOrAliasId},
     room_version_id::RoomVersionId,
     server_name::{OwnedServerName, ServerName},
@@ -218,6 +218,9 @@ pub mod __private_macros {
 
         pub static ROOM_ALIAS_ID_INTERNER: LazyLock<IdInterner<crate::RoomAliasId>> =
             LazyLock::new(IdInterner::new);
+
+        pub static ROOM_ID_INTERNER: LazyLock<IdInterner<crate::RoomId>> =
+            LazyLock::new(IdInterner::new);
     }
 }
 
@@ -279,11 +282,20 @@ macro_rules! room_id {
     };
 }
 
-/// Compile-time checked [`OwnedRoomId`] construction.
+/// Compile-time checked `&'static RoomId` construction.
+///
+/// This macro is a helper to ease the transition after the change of [`RoomId`] from a
+/// dynamically sized type to an owned type. It has the side effect of interning and leaking the
+/// identifier so it SHOULD NOT be used in code that runs in production.
+///
+/// This is behind the `unstable-identifier-ref-macros` cargo feature to allow us to remove this
+/// macro at any time without it being a breaking change.
 #[macro_export]
-macro_rules! owned_room_id {
+#[cfg(feature = "unstable-identifier-ref-macros")]
+macro_rules! room_id_ref {
     ($s:literal) => {
-        $crate::room_id!($s).to_owned()
+        $crate::__private_macros::id_interner::ROOM_ID_INTERNER
+            .get_or_insert_with($s, || $crate::room_id!($s))
     };
 }