+++ crates/ruma-federation-api/src/query/get_room_information.rs
@@ -8,7 +8,7 @@ pub mod v1 {
     //! [spec]: https://spec.matrix.org/v1.19/server-server-api/#get_matrixfederationv1querydirectory
 
     use ruma_common::{
-        OwnedRoomId, OwnedServerName, RoomAliasId,
+        OwnedServerName, RoomAliasId, RoomId,
         api::{request, response},
         metadata,
     };
@@ -34,7 +34,7 @@ pub mod v1 {
     #[response]
     pub struct Response {
         /// Room ID mapped to queried alias.
-        pub room_id: OwnedRoomId,
+        pub room_id: RoomId,
 
         /// An array of server names that are likely to hold the given room.
         pub servers: Vec<OwnedServerName>,
@@ -49,7 +49,7 @@ pub mod v1 {
 
     impl Response {
         /// Creates a new `Response` with the given room IDs and servers.
-        pub fn new(room_id: OwnedRoomId, servers: Vec<OwnedServerName>) -> Self {
+        pub fn new(room_id: RoomId, servers: Vec<OwnedServerName>) -> Self {
             Self { room_id, servers }
         }
     }