+++ crates/ruma-macros/src/lib.rs
@@ -548,7 +548,7 @@ pub fn room_alias_id(input: TokenStream) -> TokenStream {
 #[proc_macro]
 pub fn room_id(input: TokenStream) -> TokenStream {
     let id_ctor = parse_macro_input!(input as IdentifierConstructor);
-    id_ctor.validate_and_expand_str_conversion("&RoomId", room_id::validate).into()
+    id_ctor.validate_and_expand_str_conversion("RoomId", room_id::validate).into()
 }
 
 /// Compile-time checked `RoomVersionId` construction.